feat : 탈퇴구현

+++ src/main/java/eum/backed/server/service/community/ApplyService.java
@@ -7,6 +7,8 @@ import eum.backed.server.controller.community.dto.request.enums.MarketType;
 import eum.backed.server.controller.community.dto.response.ApplyResponseDTO;
 import eum.backed.server.domain.community.apply.Apply;
 import eum.backed.server.domain.community.apply.ApplyRepository;
+import eum.backed.server.domain.community.chat.ChatRoom;
+import eum.backed.server.domain.community.chat.ChatRoomRepository;
 import eum.backed.server.domain.community.marketpost.MarketPost;
 import eum.backed.server.domain.community.marketpost.MarketPostRepository;
 import eum.backed.server.domain.community.marketpost.Status;
@@ -32,10 +34,13 @@ public class ApplyService {
     private final ApplyResponseDTO applyResponseDTO;
     private final ChatService chatService;
 
+    private final ChatRoomRepository chatRoomRepository;
+
 
     public APIResponse doApply(Long postId,ApplyRequestDTO.Apply applyRequest, String email) {
         Users getUser = usersRepository.findByEmail(email).orElseThrow(() -> new NullPointerException("Invalid email"));
         MarketPost getMarketPost = marketPostRepository.findById(postId).orElseThrow(() -> new NullPointerException("Invalid postId"));
+        if(getMarketPost.isDeleted()) throw new IllegalArgumentException("Deleted post");
         if(getMarketPost.getUser() == getUser) throw new IllegalArgumentException("자기 게시글에는 신청할수 없습니다");
         if(getMarketPost.getMarketType()== MarketType.PROVIDE_HELP && getMarketPost.getPay() > getUser.getUserBankAccount().getBalance())
             throw new IllegalArgumentException("잔액보다 큰 요구 햇살");
@@ -51,7 +56,7 @@ public class ApplyService {
     public APIResponse<List<ApplyResponseDTO.ApplyListResponse>> getApplyList(Long postId) {
 //        Users getUser = usersRepository.findByEmail(email).orElseThrow(() -> new NullPointerException("Invalid email"));
         MarketPost getMarketPost = marketPostRepository.findById(postId).orElseThrow(() -> new NullPointerException("Invalid id"));
-//        List<MarketPost> marketPosts = marketPostRepository.findByUserOrderByCreateDateDesc(getUser).orElse(Collections.emptyList()); //로그인 유저가 작성한 게시글 목록 조회
+//        List<MarketPost> marketPosts = marketPostRepository.findByUserAndIsDeletedFalseOrderByCreateDateDesc(getUser).orElse(Collections.emptyList()); //로그인 유저가 작성한 게시글 목록 조회
         List<ApplyResponseDTO.ApplyListResponse> getAllApplicants = findByTransactionPosts(getMarketPost);
         return APIResponse.of(SuccessCode.SELECT_SUCCESS, getAllApplicants);
     }
@@ -72,9 +77,10 @@ public class ApplyService {
         applyIds.stream().forEach(applyId -> {
             Apply getApply = applyRepository.findById(applyId).orElseThrow(() -> new NullPointerException("invalid applyId"));
             if (getApply.getMarketPost().getUser() != getUser) throw new IllegalArgumentException("해당 게시글에 대한 권한이 없다");
-            if(getApply.getIsAccepted() == true) throw new IllegalArgumentException("이미 선정한 사람입니다");
+            if(getApply.getIsAccepted() == true || getApply.getStatus() == eum.backed.server.domain.community.apply.Status.TRADING_CANCEL) throw new IllegalArgumentException("이미 선정했더나 과거 거래 취소를 했던 사람입니다");
             MarketPost marketPost = getApply.getMarketPost();
             getApply.updateAccepted(true);
+            getApply.updateStatus(eum.backed.server.domain.community.apply.Status.TRADING);
             marketPost.addCurrentAcceptedPeople();
             if(marketPost.getCurrentAcceptedPeople() == marketPost.getMaxNumOfPeople()) {
                 marketPost.updateStatus(Status.RECRUITMENT_COMPLETED);
@@ -95,9 +101,30 @@ public class ApplyService {
     public APIResponse unApply(Long postId, Long applyId, String email) {
         Users getUser = usersRepository.findByEmail(email). orElseThrow(() -> new NullPointerException("Invalid email"));
         Apply getApply = applyRepository.findById(applyId).orElseThrow(() -> new NullPointerException("invalid applyId"));
+        if(getApply.getMarketPost().getMarketPostId() != postId) throw new IllegalArgumentException("invalid postId");
         if(getApply.getUser() != getUser) throw new IllegalArgumentException("신청 취소할 권한이 없습니다");
         if(getApply.getIsAccepted() == true) throw new IllegalArgumentException("이미 선정되서 취소할 수 없습니다");
         applyRepository.delete(getApply);
+        return APIResponse.of(SuccessCode.DELETE_SUCCESS);
+
+    }
+
+    public APIResponse cancel(Long postId, Long chatId, String email) {
+        Users getUser = usersRepository.findByEmail(email). orElseThrow(() -> new NullPointerException("Invalid email"));
+        ChatRoom chatRoom = chatRoomRepository.findById(chatId).orElseThrow(() -> new IllegalArgumentException("해딩 채팅방이 없습니다"));
+        if(chatRoom.getMarketPost().getMarketPostId() != postId) throw new IllegalArgumentException("invalid postId");
+        MarketPost getMarketPost = chatRoom.getMarketPost();
+        Apply getApply = applyRepository.findByUserAndMarketPost(chatRoom.getApplicant(), getMarketPost).orElseThrow(()->new IllegalArgumentException("신청한 이력이 없는데 채팅방이 있다"));
+        if(!(getUser == chatRoom.getApplicant() || getUser == chatRoom.getPostWriter() )) throw new IllegalArgumentException("활동 파기할수있는 권한이 없습니다");
+//        블록 상태 처리
+        chatRoom.upDateBlocked(true);
+        getApply.updateStatus(eum.backed.server.domain.community.apply.Status.TRADING_CANCEL);
+        getApply.updateAccepted(false);
+        applyRepository.save(getApply);
+
+        getMarketPost.subCurrentAcceptedPeople();
+        getMarketPost.updateStatus(Status.RECRUITING);
+        marketPostRepository.save(getMarketPost);
         return APIResponse.of(SuccessCode.DELETE_SUCCESS);
 
     }
